requirements.txt
ujson
nltk
pytorch
scikit-learn
gensim
py-rouge
pyrouge
matplotlib
tensorboardX
pyyaml
tqdm
